Added start() to ::Session
The start() method takes the fh of the new connection and starts a
AnyEvent::Handle object with it.

We use this for all buffered I/O with the client.

For now, we just wait for EOF or errors and disconnect, clearing the
::Server session entry.

A new attribute, `handle`, was added to ::Session to keep track of the
ANyEvent::Handle object we create.

Signed-off-by: Pedro Melo <melo@simplicidade.org>