Update the server to use Async::Hooks for all command processing
 * removed old ::Parser: most code had moved to ::Utils;
 * added ::Commands for all SMTP command parsing, validation and
   execution: for now it supports HELO, EHLO and MAIL;
 * all objects created by ::Session have a slot for class name: this
   allows other developers to override all the classes the server uses;
 * ::Session creates a ::Request for each input line in command mode: this
   request is passed along on all callbacks and is used to store the
   parsed elements of the command;
 * update ::Transaction->reverse_path to use our new ::Path objects: allows
   us to store the address and all the extensions used.

The current parse, validation and execution sequence of events is the
following.

Foreach line read from the client, the server calls the 'line_in' hook.
If you want to override the entire processing of the line or if you
just want to log the entire session, this is the place to hook into.

The default action is to parse the SMTP command, spliting the line in two,
a command and the rest.

With the command name (for example CMD), we check the existence of a chain
named 'parse_CMD_command'. If no such chain exists, we reply with an
invalid command error.

If the chain exists, we call it to parse the command. This chain must
parse the rest of the arguments and fill the appropriate request
object slots.

If the entire line is parsed, the chain 'validate_CMD_command'
is called next.

This allows extensions to be checked with the full request already parsed.
You must use the ack_extensions() method of ::Request to acknowledge your
own SMTP extensions if any.

The default action of this chain will check if there are any
unacknowledged extensions and report a syntax error if any.

Finally, if everything was ok so far, the chain 'execute_CMD_command' is called.

The default action is to report a generic command error, so one
callback on this chain must generate the proper response and
call $ctl->done to signal that the command was processed.

Signed-off-by: Pedro Melo <melo@simplicidade.org>